[wip] 11-2 + HashMap impl

+++ 2024/libs/include/array_list.h
@@ -5,18 +5,18 @@
 #include <stdint.h>
 
 typedef struct ArrayList {
-  uint32_t count;
-  uint32_t capacity;
-  uint32_t *head;
+  uint64_t count;
+  uint64_t capacity;
+  uint64_t *head;
 } ArrayList;
 
-ArrayList *array_list_init(const uint32_t capacity);
+ArrayList *array_list_init(const uint64_t capacity);
 
-void array_list_add(ArrayList *this, const uint32_t value);
+void array_list_add(ArrayList *this, const uint64_t value);
 
-uint32_t array_list_get(const ArrayList *this, const uint32_t index);
+uint64_t array_list_get(const ArrayList *this, const uint64_t index);
 
-bool array_list_contains(const ArrayList *this, const uint32_t value);
+bool array_list_contains(const ArrayList *this, const uint64_t value);
 
 void array_list_clear(ArrayList *this);
